backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.15
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-grappelli==3.0.4
django-jet-reboot==1.3.3
django-redis==5.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
